Apply changes
